requirements.txt
langchain==1.1.2
langchain-core==1.1.1
langchain-openai==1.1.0
langgraph>=0.0.30
anthropic>=0.18.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
newsapi-python>=0.2.7
praw>=7.7.1
# redis is optional for MVP
redis>=5.0.0
pytest
pytest-mock
# Observability & Tracing
# arize-phoenix>=4.0.0  # Not compatible with Python 3.14 yet
# openinference-instrumentation-langchain>=0.1.0  # Depends on arize-phoenix
# Web Interface
streamlit>=1.28.0
